feat: Improve validation of poll proposal quorum and approval upon updating priority
